+++ engine/src/dpx_jpegimage.rs
@@ -444,8 +444,8 @@ unsafe extern "C" fn jpeg_get_density(
 unsafe extern "C" fn JPEG_info_init(mut j_info: *mut JPEG_info) {
     (*j_info).width = 0_u16;
     (*j_info).height = 0_u16;
-    (*j_info).bits_per_component = 0i32 as u8;
-    (*j_info).num_components = 0i32 as u8;
+    (*j_info).bits_per_component = 0_u8;
+    (*j_info).num_components = 0_u8;
     (*j_info).xdpi = 0.0f64;
     (*j_info).ydpi = 0.0f64;
     (*j_info).flags = 0i32;
@@ -715,10 +715,10 @@ unsafe extern "C" fn read_APP1_Exif(
         if !(p.offset(8) >= endptr) {
             tiff_header = p;
             if *p as i32 == 'M' as i32 && *p.offset(1) as i32 == 'M' as i32 {
-                bigendian = 0i32 as i8;
+                bigendian = 0_i8;
                 current_block = 1109700713171191020;
             } else if *p as i32 == 'I' as i32 && *p.offset(1) as i32 == 'I' as i32 {
-                bigendian = 1i32 as i8;
+                bigendian = 1_i8;
                 current_block = 1109700713171191020;
             } else {
                 dpx_warning(
